MGPU shadow first pass Added 1/8 compression for shadow masks

+++ GraphicsEngine/Source/Core/Rendering/RenderNodes/StoreNodes/BufferStorageNode.h
@@ -15,7 +15,7 @@ public:
 		return GPUBuffer;
 	};
 	RHIBufferDesc Desc = RHIBufferDesc();
-	int LinkedFrameBufferRatio = 1;
+	float LinkedFrameBufferRatio = 1.0f;
 	FrameBufferStorageNode* FramebufferNode = nullptr;
 protected:
 	void Create() override;
